merge

+++ src/OrchardCore.Modules/OrchardCore.Lucene/Services/LuceneIndexingService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using OrchardCore.ContentLocalization;
 using OrchardCore.ContentManagement;
 using OrchardCore.Entities;
 using OrchardCore.Environment.Shell;
@@ -186,7 +187,9 @@ namespace OrchardCore.Lucene
                                     continue;
                                 }
 
-                                var ignoreIndexedCulture = settings.Culture == "any" ? false : context.ContentItem.Content?.LocalizationPart?.Culture != settings.Culture;
+                                var cultureAspect = await contentManager.PopulateAspectAsync<CultureAspect>(context.ContentItem);
+                                var culture = cultureAspect.HasCulture ? cultureAspect.Culture.Name : null;
+                                var ignoreIndexedCulture = settings.Culture == "any" ? false : culture != settings.Culture;
 
                                 // Ignore if the content item content type or culture is not indexed in this index
                                 if (!settings.IndexedContentTypes.Contains(context.ContentItem.ContentType) || ignoreIndexedCulture)
@@ -225,7 +228,7 @@ namespace OrchardCore.Lucene
                     }
 
                     _indexingState.Update();
-                });
+                }, activateShell: false);
             } while (batch.Length == BatchSize);
         }
 
@@ -254,7 +257,11 @@ namespace OrchardCore.Lucene
         /// <returns></returns>
         public Task DeleteIndexAsync(string indexName)
         {
-            _indexManager.DeleteIndex(indexName);
+            if (_indexManager.Exists(indexName))
+            {
+                _indexManager.DeleteIndex(indexName);
+            }
+
             return _luceneIndexSettingsService.DeleteIndexAsync(indexName);
         }
 
@@ -273,7 +280,11 @@ namespace OrchardCore.Lucene
         /// </summary>
         public async Task RebuildIndexAsync(string indexName)
         {
-            _indexManager.DeleteIndex(indexName);
+            if (_indexManager.Exists(indexName))
+            {
+                _indexManager.DeleteIndex(indexName);
+            }
+
             await _indexManager.CreateIndexAsync(indexName);
 
             ResetIndex(indexName);
